api_database/requirements.txt
asgiref==3.5.2
Django==3.2.23
djangorestframework==3.14.0
django-filter==22.1
drf-yasg==1.21.4
numpy>=1.25.2
gemmi==0.5.8
djoser==2.1.0
networkx==2.8.8
PyCifRW==4.4.5
